MSc Professor Changes Request
Implemented the changes as instructed by the MSc Professor overseeing the project. Implemented a sources section with the ability to redirect to the respective sources in this cage the two kaggle datasets utilizing their urls. Added an Author section with our names on the footer. Also as an extra added a "back to the top" button for easier navigation.